+++ vote.c
@@ -193,7 +193,7 @@ unsigned char *tmpsave;
 
 
   {
-    int size = sizeof (vote.msg) - 1;
+    size_t size = sizeof (vote.msg) - 1;
     if (!(tmpstart = (unsigned char *)mymmap(NULL, &size, 1)))
       return(MMAP_ERR);
   }
